% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_shapefile.R
\name{export_shapefile}
\alias{export_shapefile}
\title{Get landscape summary}
\usage{
export_shapefile(
  raster_path,
  raster_value,
  shapefile_path,
  neighborhoodExtractCC = 4,
  neighborhoodExploreCC = 4
)
}
\description{
Get landscape summary
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
