% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_intra_metrics.R
\name{extract_intra_metrics}
\alias{extract_intra_metrics}
\title{Compute all intra-patch connectivity metric.}
\usage{
extract_intra_metrics(
  raster_path,
  raster_value,
  neighborhoodExtractCC = 4,
  neighborhoodExploreCC = 4
)
}
\value{
A dataframe
}
\description{
Create a new restoration optimization problem using data that describe
the spatial distribution of existing habitat and potential for
restoration.
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
