% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_mesh_cwa.R
\name{intra_mesh}
\alias{intra_mesh}
\title{Compute MESH}
\usage{
intra_mesh(intra_metrics, landscape_summary)
}
\arguments{
\item{intra_metrics}{dataframe obtained using `extract_intra_metrics`}
}
\value{
A dataframe
}
\description{
Compute MESH
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
