% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_mesh_cwa.R
\name{intra_mesh_cwa_frac}
\alias{intra_mesh_cwa_frac}
\title{Compute MESH_CWA[FRAC]}
\usage{
intra_mesh_cwa_frac(intra_metrics, landscape_summary)
}
\arguments{
\item{intra_metrics}{dataframe obtained using `extract_intra_metrics`}
}
\value{
A dataframe
}
\description{
Compute MESH_CWA[FRAC]
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
