% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_mesh_cwa.R
\name{intra_mesh_cwa_shape}
\alias{intra_mesh_cwa_shape}
\title{Compute MESH_CWA[SHAPE]}
\usage{
intra_mesh_cwa_shape(intra_metrics)
}
\arguments{
\item{intra_metrics}{dataframe obtained using `extract_intra_metrics`}
}
\value{
A dataframe
}
\description{
Compute MESH_CWA[SHAPE]
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
