% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_summary.R
\name{landscape_summary}
\alias{landscape_summary}
\title{Get landscape summary}
\usage{
landscape_summary(
  raster_path,
  raster_value,
  neighborhoodExtractCC = 4,
  neighborhoodExploreCC = 4
)
}
\value{
A dataframe
}
\description{
Get landscape summary
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
